pyproject.toml
[project]
name = "ai-server-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
